created UserTest; registerUser_validData_noExistingUser_success: This test creates two users, one of which already exists in the store (has the same username). The test checks that the new user successfully validates and is not equal to an existing user.
registerUser_shortLogin_exceptionThrown: This test creates a user with a login that is too short (less than 6 characters). The test verifies that the user throws a custom unchecked InvalidDataException during validation.

registerUser_shortPassword_exceptionThrown: This test creates a user with a password that is too short (less than 6 characters). The test verifies that the user throws a custom unchecked InvalidDataException during validation.

registerUser_underageUser_exceptionThrown: This test creates an underage user (under age 18). The test verifies that the user throws a custom unchecked InvalidDataException during validation.
registerUser_nullLogin_exceptionThrown: Checks that attempting to register a user with a null login throws an InvalidDataException.
registerUser_nullPassword_exceptionThrown: Checks that attempting to register a user with a null password throws an InvalidDataException.
registerUser_negativeAge_exceptionThrown: Checks that attempting to register a user with a negative age will throw an InvalidDataException.
registerUser_upperAgeLimit_validData_success: Checks that attempting to register a user with the maximum allowed age value (eg 150) does not throw an exception and the validation succeeds..